refactor post tags
